+++ neovim.cpp/include/socket.hpp
@@ -26,13 +26,13 @@ public:
     }
 
     void connect_tcp(const std::string& host,
-               const std::string& service, double timeout_sec);
+               const std::string& service, long timeout_sec);
 
-    size_t read(char *rbuf, size_t capacity, double timeout_microsec);
+    size_t read(char *rbuf, size_t capacity, long timeout_microsec);
 
-    size_t ui_read(char *rbuf, size_t capacity, double timeout_microsec);
+    size_t ui_read(char *rbuf, size_t capacity, long timeout_microsec);
 
-    void write(char *sbuf, size_t size, double timeout_sec);
+    void write(char *sbuf, size_t size, long timeout_sec);
 
     size_t available(){ return socket_.available(); }
